Adds support for multiple signers on direct jobs
Consumers of the client may specify multiple signers on a direct signature job.
After creation the service will return one redirect URL for each signer to where the sender will redirect the user.
When receiving an updated status the service will return an aggregated status for the whole job in addition to a status for each signer.
The signer statuses are changed from an enum to a collection of constants because there may be additional statuses which are not known to older versions of the client library in the future.
A new test pattern has been introduced for the `DirectClientSmokeTest`s in order to improve readability. Lets try it out for a while and rewrite more tests to follow this pattern if it feels good.