Remove Get_pades from direct multi signer-test
Because of the way the job is signed (using the autosign backdoor) we won't be able to retrieve a PAdES document from Signicat.
Retrieving PAdES for a direct job is verified working in the single signer-test, so this is most likely not a problem.
If we really, really want to test the Get_pades()-method we need to find a better way of handling XAdESes in the back end.