tests: Disable client side certificate validation
We have updated the client test certificates to internally issued ones
instead of Buypass certificates. This should not affect any environments
as our cert authority should be configured but the client side
validation in this code base is strict by default.

Disable the validation locally. Server side validation is still active
so using a certificate not originating from a valid authority or
Digipost will fail.