Ensure non-sensitive title is passed onto manifest.
Manifest is part of document bundle sent to the server.

A Portal.Job object might be newed up with a non-sensitive title, but
this was lost in the mapping from Portal.Job to Manifest. This resulted
in a bug which made it impossible for consumers of our dotnet client
library to create portal jobs with non-sensitive tiles.

Non-sensitive titles are titles that may be added in addition to
"regular" titles. Non-sensitive titles may be displayed in contexts
where we cannot ensure that people viewing are authorized, such as in
email notificaitons sent to signers.